feat: enhance RpcMessage and EventListener types (#211)
* feat: enhance RpcMessage and EventListener types

* chore: export more types
